+++ src/gsExpressions/divide_expr.h
@@ -1,0 +1,137 @@
+/** @file divide_expr.h
+
+    @brief Defines the divide expression
+
+    This file is part of the G+Smo library.
+
+    This Source Code Form is subject to the terms of the Mozilla Public
+    License, v. 2.0. If a copy of the MPL was not distributed with this
+    file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+    Author(s): A. Mantzaflaris
+               H.M. Verhelst
+*/
+
+#pragma once
+
+namespace gismo
+{
+namespace expr
+{
+
+/*
+  Expression for scalar division operation (first version)
+*/
+template <typename E1, typename E2>
+class divide_expr : public _expr<divide_expr<E1,E2> >
+{
+    typename E1::Nested_t _u;
+    typename E2::Nested_t _v;
+
+public:
+    typedef typename E1::Scalar Scalar;
+
+public:
+    enum {ScalarValued = E1::ScalarValued, ColBlocks= E2::ColBlocks};
+    enum {Space = E1::Space}; // The denominator E2 has to be scalar.
+
+    divide_expr(_expr<E1> const& u, _expr<E2> const& v)
+    : _u(u), _v(v)
+    {
+        GISMO_STATIC_ASSERT(E2::ScalarValued, "The denominator needs to be scalar valued.");
+    }
+
+    AutoReturn_t eval(const index_t k) const
+    { return ( _u.eval(k) / _v.eval(k) ); }
+
+    index_t rows() const { return _u.rows(); }
+    index_t cols() const { return _u.cols(); }
+
+    void parse(gsExprHelper<Scalar> & evList) const
+    { _u.parse(evList); _v.parse(evList); }
+
+
+    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
+    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
+
+    void print(std::ostream &os) const
+    { os << "("; _u.print(os);os <<" / ";_v.print(os);os << ")"; }
+};
+
+/*
+  Division specialization (second version) for constant value denominator
+*/
+template <typename E1>
+class divide_expr<E1,typename E1::Scalar>
+    : public _expr<divide_expr<E1,typename E1::Scalar> >
+{
+public:
+    typedef typename E1::Scalar Scalar;
+
+private:
+    typename E1::Nested_t _u;
+    Scalar  const   _c;
+
+public:
+    enum {Space= E1::Space, ScalarValued = E1::ScalarValued, ColBlocks= E1::ColBlocks};
+
+    divide_expr(_expr<E1> const& u, Scalar const  c)
+    : _u(u), _c(c) { }
+
+    AutoReturn_t eval(const index_t k) const
+    { return ( _u.eval(k) / _c ); }
+
+    index_t rows() const { return _u.rows(); }
+    index_t cols() const { return _u.cols(); }
+
+    void parse(gsExprHelper<Scalar> & evList) const
+    { _u.parse(evList); }
+
+
+    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
+    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
+
+    void print(std::ostream &os) const
+    { os << "("; _u.print(os);os <<"/"<< _c << ")"; }
+};
+
+/*
+  Division specialization (third version) for constant value
+  numerator
+*/
+template <typename E2>
+class divide_expr<typename E2::Scalar,E2>
+    : public _expr<divide_expr<typename E2::Scalar,E2> >
+{
+public:
+    typedef typename E2::Scalar Scalar;
+
+private:
+    Scalar  const   _c;
+    typename E2::Nested_t _u;
+public:
+    enum {Space= 0, ScalarValued = 1, ColBlocks= 0};
+
+    divide_expr(Scalar const c, _expr<E2> const& u)
+    : _c(c), _u(u)
+    { GISMO_STATIC_ASSERT(E2::ScalarValued, "The denominator needs to be scalar valued."); }
+
+    Scalar eval(const index_t k) const
+    { return ( _c / _u.val().eval(k) ); }
+
+    index_t rows() const { return 0; }
+    index_t cols() const { return 0; }
+
+    void parse(gsExprHelper<Scalar> & evList) const
+    { _u.parse(evList); }
+
+
+    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
+    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
+
+    void print(std::ostream &os) const
+    { os << "("<< _c <<"/";_u.print(os);os << ")";}
+};
+
+}// namespace expr
+}// namespace gismo
