+++ src/gsExpressions/gsExpressions.h
@@ -101,8 +101,8 @@ template<class E> class colsum_expr;
 template<class E> class col_expr;
 template<class T> class meas_expr;
 template<class E> class inv_expr;
-template<class E, bool cw = false> class tr_expr;
-template<class E> class cb_expr;
+template<class E, bool cw = false> class transpose_expr;
+template<class E> class colBlocks_expr;
 template<class E> class abs_expr;
 template<class E> class pow_expr;
 template<class E> class sign_expr;
@@ -201,16 +201,16 @@ public:
     { return static_cast<E const&>(*this).eval(k); }
 
     /// Returns the transpose of the expression
-    tr_expr<E> tr() const
-    { return tr_expr<E,false>(static_cast<E const&>(*this)); }
+    transpose_expr<E> tr() const
+    { return transpose_expr<E,false>(static_cast<E const&>(*this)); }
 
     /// Returns the coordinate-wise transpose of the expression
-    tr_expr<E,true> cwisetr() const
-    { return tr_expr<E,true>(static_cast<E const&>(*this)); }
+    transpose_expr<E,true> cwisetr() const
+    { return transpose_expr<E,true>(static_cast<E const&>(*this)); }
 
     /// Returns the puts the expression to colBlocks
-    cb_expr<E> cb() const
-    { return cb_expr<E>(static_cast<E const&>(*this)); }
+    colBlocks_expr<E> cb() const
+    { return colBlocks_expr<E>(static_cast<E const&>(*this)); }
 
     /// Returns the sign of the expression
     sign_expr<E> sgn(Scalar tolerance=0) const
@@ -368,9 +368,78 @@ std::ostream &operator<<(std::ostream &os, const _expr<E> & b)
 #include <gsExpressions/gsGeometryMap.h>
 
 #include <gsExpressions/symbol_expr.h>
-#include <gsExpressions/integral_expr.h>
+
+
+// A
+#include <gsExpressions/abs_expr.h>
+#include <gsExpressions/add_expr.h>
+#include <gsExpressions/adjugate_expr.h>
+#include <gsExpressions/asdiag_expr.h>
+// B
+// C
 #include <gsExpressions/col_expr.h>
+#include <gsExpressions/colsum_expr.h>
+#include <gsExpressions/constMat_expr.h>
+#include <gsExpressions/colBlocks_expr.h>
+#include <gsExpressions/curl_expr.h>
+// D
+#include <gsExpressions/diag_expr.h>
+// #include <gsExpressions/div_expr.h> //todo
+#include <gsExpressions/divide_expr.h>
+// E
+#include <gsExpressions/exp_expr.h>
+// F
+#include <gsExpressions/flat_expr.h>
+#include <gsExpressions/frprod_expr.h>
+// G
+#include <gsExpressions/grad_expr.h>
+// H
+// I
+#include <gsExpressions/idMat_expr.h>
+#include <gsExpressions/integral_expr.h>
+// J
+// K
+// L
+#include <gsExpressions/lapl_expr.h>
+// M
+#include <gsExpressions/max_expr.h>
 #include <gsExpressions/meas_expr.h>
+#include <gsExpressions/mult_expr.h>
+// N
+#include <gsExpressions/nabla_expr.h>
+#include <gsExpressions/nabla2_expr.h>
+#include <gsExpressions/normal_expr.h>
+// O
+#include <gsExpressions/onormal_expr.h>
+// P
+#include <gsExpressions/pow_expr.h>
+// Q
+// R
+#include <gsExpressions/reshape_expr.h>
+#include <gsExpressions/replicate_expr.h>
+#include <gsExpressions/rowsum_expr.h>
+// S
+#include <gsExpressions/sign_expr.h>
+#include <gsExpressions/sub_expr.h>
+// T
+#include <gsExpressions/tangent_expr.h>
+#include <gsExpressions/temp_expr.h>
+#include <gsExpressions/trace_expr.h>
+#include <gsExpressions/transpose_expr.h>
+// U
+// V
+#include <gsExpressions/value_expr.h>
+#include <gsExpressions/voigt_expr.h> // @hverhelst todo: add and replace flat_expr
+// W
+// X
+// Y
+// Z
+
+
+
+
+
+
 
 namespace gismo
 {
@@ -406,484 +475,6 @@ public:
     void print(std::ostream &os) const { os<<_c; }
 };
 
-
-
-
-/*
-  Expression for the transpose of an expression
-*/
-template<class E, bool cw>
-class tr_expr : public _expr<tr_expr<E,cw> >
-{
-    typename E::Nested_t _u;
-
-public:
-
-    typedef typename E::Scalar Scalar;
-
-    tr_expr(_expr<E> const& u)
-    : _u(u) { }
-
-public:
-    enum {ColBlocks = E::ColBlocks, ScalarValued=E::ScalarValued};
-    enum {Space = cw?E::Space:(E::Space==1?2:(E::Space==2?1:E::Space))};
-
-    mutable Temporary_t res;
-    const Temporary_t & eval(const index_t k) const
-    {
-        if (E::ColBlocks)
-            res = _u.eval(k).blockTranspose( _u.cardinality() );
-        else
-            res = _u.eval(k).transpose();
-        return res;
-    }
-
-    index_t rows() const { return _u.cols(); }
-
-    index_t cols() const { return _u.rows(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return cw?_u.rowVar():_u.colVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return cw?_u.colVar():_u.rowVar(); }
-
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void print(std::ostream &os) const { os<<"("; _u.print(os); os <<")\u1D40"; }
-private:
-/*
-  template<class U> EIGEN_STRONG_INLINE MatExprType
-  eval_impl(const U k, typename util::enable_if<1==ColBlocks,U>::type* = nullptr)
-  { return _u.eval(k).blockTranspose(_u.cols()/_u.rows()); }
-
-  template<class U> EIGEN_STRONG_INLINE MatExprType
-  eval_impl(const U k, typename util::enable_if<0==ColBlocks,U>::type* = nullptr)
-  { return _u.eval(k).transpose(); }
-*/
-};
-
-/*
-  Expression to make an expression colblocks
-*/
-template<class E>
-class cb_expr : public _expr<cb_expr<E> >
-{
-    typename E::Nested_t _u;
-
-public:
-
-    typedef typename E::Scalar Scalar;
-
-    cb_expr(_expr<E> const& u)
-    : _u(u) { }
-
-public:
-    enum {ColBlocks = 1, ScalarValued=E::ScalarValued};
-    enum {Space = E::Space};
-
-    mutable gsMatrix<Scalar> ev, res;
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        //return _u.eval(k).transpose();
-        // /*
-        ev = _u.eval(k);
-        if (E::ColBlocks)
-        {
-            return ev;
-        }
-        else
-        {
-            res = _u.eval(k);
-
-            GISMO_ASSERT(res.rows() % _u.rows() == 0 && res.cols() % _u.cols() == 0,"Result is not a multiple of the space dimensions?");
-
-            index_t cardinality;
-            if ( (cardinality = res.rows() / _u.rows()) >= 1 && res.cols() / _u.cols() == 1 ) // stored in rows
-            {
-                res.resize(_u.rows(), cardinality * _u.cols());
-                for (index_t r = 0; r!=cardinality; r++)
-                    res.block(0 , r * _u.cols(), _u.rows(), _u.cols()) = ev.block( r * _u.rows(), 0, _u.rows(), _u.cols() );
-            }
-            else if ( (cardinality = res.rows() / _u.rows()) == 1 && res.cols() / _u.cols() >= 1 ) // stored in cols ----->>>> This is already colBlocks???
-            {
-                res.resize(_u.rows(), cardinality * _u.cols());
-                for (index_t r = 0; r!=cardinality; r++)
-                    res.block(0 , r * _u.cols(), _u.rows(), _u.cols()) = ev.block( 0, r * _u.cols(), _u.rows(), _u.cols() );
-            }
-        }
-        return res;
-    }
-
-    index_t rows() const { return _u.rows(); }
-
-    index_t cols() const { return _u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    index_t cardinality_impl() const
-    {
-        res = _u.eval(0);
-        index_t cardinality;
-        if ( res.rows() / _u.rows() >= 1 && res.cols() / _u.cols() == 1 ) // stored in rows
-            cardinality = res.rows() / _u.rows();
-        else if ( res.rows() / _u.rows() == 1 && res.cols() / _u.cols() >= 1 )
-            cardinality = res.cols() / _u.cols();
-        else
-            GISMO_ERROR("Cardinality for cb_expr cannot be determined.");
-
-        return cardinality;
-    }
-
-    void print(std::ostream &os) const { os<<"{"; _u.print(os); os <<"}"; }
-};
-
-
-/*
-  Expression for an evaluation of the (sub-)expression in temporary memory
-*/
-template<class E>
-class temp_expr : public _expr<temp_expr<E> >
-{
-    typename E::Nested_t _u;
-    typedef typename E::Scalar Scalar;
-    mutable gsMatrix<Scalar> tmp;
-
-public:
-    temp_expr(_expr<E> const& u)
-    : _u(u) { }
-
-public:
-    enum {Space = E::Space, ScalarValued = E::ScalarValued,
-        ColBlocks = E::ColBlocks};
-
-    // template<bool S  = ColBlocks>
-    // typename util::enable_if<S,MatExprType>::type
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        tmp = _u.eval(k);
-        return tmp;
-    }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-    void parse(gsExprHelper<Scalar> & evList) const { _u.parse(evList); }
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const { _u.print(os); }
-};
-
-/*
-  Expression for the trace of a (matrix) expression
-*/
-template<class E>
-class trace_expr  : public _expr<trace_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E::Space, ColBlocks= 0};
-
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> res;
-
-public:
-    trace_expr(_expr<E> const& u) : _u(u)
-    {
-        // gcc 4.8.4: invalid read due to _u.rows() using gsFuncData
-        //GISMO_ASSERT(0== _u.cols()%_u.rows(), "Expecting square-block expression, got " << _u.rows() <<" x "<< _u.cols() );
-    }
-
-    // choose if ColBlocks
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        auto tmp = _u.eval(k);
-        const index_t cb = _u.rows();
-        const index_t r  = _u.cardinality();
-        if (Space==1)
-            res.resize(r, 1);
-        else
-            res.resize(1, r);
-
-        for (index_t i = 0; i!=r; ++i)
-            res.at(i) = tmp.middleCols(i*cb,cb).trace();
-        return res;
-    }
-
-    // choose if !ColBlocks
-    //todo: Scalar eval(const index_t k) const
-
-    index_t rows() const { return _u.cols() / _u.rows(); } //_u.cardinality()?
-    index_t cols() const { return 1; }
-
-    index_t cardinality_impl() const { return _u.cardinality(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const { os << "trace("; _u.print(os); os<<")"; }
-};
-
-/*
-  Expression for the adjugate of a (matrix) expression
-*/
-template<class E>
-class adjugate_expr  : public _expr<adjugate_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, ColBlocks = E::ColBlocks};
-    enum {Space = E::Space};
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> res;
-
-public:
-    adjugate_expr(_expr<E> const& u) : _u(u)
-    {
-        // gcc 4.8.4: invalid read due to _u.rows() using gsFuncData
-        //GISMO_ASSERT(0== _u.cols()%_u.rows(), "Expecting square-block expression, got " << _u.rows() <<" x "<< _u.cols() );
-    }
-
-    // choose if ColBlocks
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        auto tmp = _u.eval(k);
-        const index_t cb = _u.rows();
-        const index_t r  = _u.cols() / cb;
-        res.resize(_u.rows(),_u.cols());
-        for (index_t i = 0; i!=r; ++i){
-            res.middleCols(i*cb,cb) = tmp.middleCols(i*cb,cb).adjugate();
-        }
-        return res;
-    }
-
-    // choose if !ColBlocks
-    //todo: Scalar eval(const index_t k) const
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const { os << "adj("; _u.print(os); os<<")"; }
-};
-
-template<class E>
-class reshape_expr  : public _expr<reshape_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, ColBlocks = E::ColBlocks};
-    enum {Space = E::Space};
-private:
-    typename E::Nested_t _u;
-    index_t _n, _m;
-    mutable gsMatrix<Scalar> tmp;
-
-public:
-
-    //the reshaping is done column-wise
-    reshape_expr(_expr<E> const& u, index_t n, index_t m) : _u(u), _n(n), _m(m)
-    {
-        //GISMO_ASSERT( _u.rows()*_u.cols() == _n*_m, "Wrong dimension"); //
-    }
-
-    const gsAsConstMatrix<Scalar> eval(const index_t k) const
-    {
-        // Note: this assertion would fail in the constructor!
-        GISMO_ASSERT( _u.rows()*_u.cols() == _n*_m, "Wrong dimension "
-                      << _u.rows() << " x "<<_u.cols() << "!=" << _n << " * "<< _m );
-        tmp = _u.eval(k);
-        return gsAsConstMatrix<Scalar>(tmp.data(),_n,_m);
-    }
-
-    index_t rows() const { return _n; }
-    index_t cols() const { return _m; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const { os << "reshape("; _u.print(os); os<<","<<_n<<","<<_m<<")"; }
-};
-
-/// Reshape an expression
-template <typename E> EIGEN_STRONG_INLINE
-reshape_expr<E> const reshape(E const & u, index_t n, index_t m)
-{ return reshape_expr<E>(u, n, m); }
-
-template<class E>
-class replicate_expr  : public _expr<replicate_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E::Space, ColBlocks= E::ColBlocks};
-private:
-    typename E::Nested_t _u;
-    index_t _n, _m;
-    mutable gsMatrix<Scalar> tmp;
-
-public:
-
-    //the replicate is done nxm times
-    replicate_expr(_expr<E> const& u, index_t n, index_t m) : _u(u), _n(n), _m(m)
-    {
-    }
-
-    auto eval(const index_t k) const -> decltype(tmp.replicate(_n,_m))
-    {
-        tmp = _u.eval(k);
-        return tmp.replicate(_n,_m);
-    }
-
-    index_t rows() const { return _n*_u.rows(); }
-    index_t cols() const { return _m*_u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void print(std::ostream &os) const { os << "replicate("; _u.print(os); os<<","<<_n<<","<<_m<<")"; }
-};
-
-/// Replicate an expression
-template <typename E> EIGEN_STRONG_INLINE
-replicate_expr<E> const replicate(E const & u, index_t n, index_t m = 1)
-{ return replicate_expr<E>(u, n, m); }
-
-/**
-   Transforms a matrix expression into a vector expression by computing the vector
-   [ a b c+d]^T
-   for each matrix block
-   [ a d ]
-   [ c b ]
-*/
-template<class E>
-class flat_expr  : public _expr<flat_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E::Space, ColBlocks= 0}; // to do: ColBlocks
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> tmp;
-
-public:
-
-    flat_expr(_expr<E> const& u) : _u(u)
-    {
-        //GISMO_ASSERT( _u.rows()*_u.cols() == _n*_m, "Wrong dimension"); //
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        tmp = _u.eval(k);
-        const index_t numActives = _u.cardinality();
-
-        for (index_t i = 0; i<numActives; ++i)
-        {
-            tmp(0,2*i+1) += tmp(1,2*i);
-            std::swap(tmp(1,2*i), tmp(1,2*i+1));
-        }
-
-        tmp.resize(4,numActives);
-        tmp.conservativeResize(3,numActives);
-
-        if ( 1==Space )
-            tmp.transposeInPlace();
-        else if (2!=Space) // if not colSpan and not rowSpan
-            tmp.transposeInPlace();
-
-        return tmp;
-    }
-
-    index_t rows() const { return 1; }
-    index_t cols() const { return 3; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void print(std::ostream &os) const { os << "flat("; _u.print(os); os<<")"; }
-};
-
-/// Make a matrix 2x2 expression "flat"
-template <typename E> EIGEN_STRONG_INLINE
-flat_expr<E> const flat(E const & u)
-{ return flat_expr<E>(u); }
-
-
-/*
-  Expression for the diagonal(s) of a (matrix) expression
-*/
-  template<class E>
-  class diag_expr  : public _expr<diag_expr<E> >
-  {
-    public:
-        typedef typename E::Scalar Scalar;
-        enum {Space=0, ColBlocks=E::ColBlocks, ScalarValued = 0};
-    private:
-        typename E::Nested_t _u;
-        mutable gsMatrix<Scalar> res;
-
-    public:
-        diag_expr(_expr<E> const& u) : _u(u)
-        {
-            GISMO_ASSERT(0== _u.cols()%_u.rows(), "Expecting square-block expression, got "
-            << _u.rows() <<" x "<< _u.cols() );
-        }
-
-        const gsMatrix<Scalar> & eval(const index_t k) const
-        {
-            // Assume mat ??
-            MatExprType tmp = _u.eval(k);
-            const index_t cb = _u.rows();
-            const index_t r  = _u.cols() / cb;
-            res.resize(r, cb);
-            for (index_t i = 0; i!=r; ++i)
-                res.row(i) = tmp.middleCols(i*cb,cb).diagonal();
-            return res;
-        }
-
-        index_t rows() const { return _u.cols() / _u.rows(); }
-        index_t cols() const { return _u.rows(); }
-
-        void parse(gsExprHelper<Scalar> & evList) const
-        { _u.parse(evList); }
-
-        const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-        const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-
-  void print(std::ostream &os) const { os << "diag("; _u.print(os); os<<")"; }
-  };
-
-/// Get diagonal elements of matrix as a vector
-template <typename E> EIGEN_STRONG_INLINE
-diag_expr<E> const diagonal(E const & u)
-{ return diag_expr<E>(u); }
 
 #define GISMO_EXPR_VECTOR_EXPRESSION(name, mname, isSv)                 \
     template<class E> class name##_##expr  : public _expr<name##_##expr<E> > { \
@@ -923,314 +514,6 @@ GISMO_EXPR_VECTOR_EXPRESSION(det,determinant,1);
 //GISMO_EXPR_VECTOR_EXPRESSION(replicate,replicate,0);
 
 #undef GISMO_EXPR_VECTOR_EXPRESSION
-
-/**
-   Expression for turning a vector into a diagonal matrix
-*/
-template<class E>
-class asdiag_expr : public _expr<asdiag_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> res;
-
-public:
-    enum{Space = E::Space, ScalarValued= 0, ColBlocks= E::ColBlocks};
-
-    asdiag_expr(_expr<E> const& u) : _u(u) { }
-
-public:
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        auto m = _u.eval(k);
-        const index_t r = m.rows();
-        const index_t c = m.cols();
-        res.resize(r,r*c);
-        for (index_t i = 0; i!=c; ++i)
-            res.middleCols(i*r,r) = m.col(i).asDiagonal();
-        return res;
-    }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.rows() * _u.cols(); }
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    void print(std::ostream &os) const { os << "diag("; _u.print(os); os <<")";}
-};
-
-// Takes the max of a vector
-template<class E>
-class max_expr  : public _expr<max_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E::Space, ColBlocks = 1};
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> tmp;
-    mutable gsMatrix<Scalar> res;
-
-public:
-
-    max_expr(_expr<E> const& u) : _u(u)
-    {
-        //GISMO_ASSERT( _u.rows()*_u.cols() == _n*_m, "Wrong dimension"); //
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const {return eval_impl(_u,k); }
-
-    index_t rows() const { return 1; }
-    index_t cols() const { return 1; }
-    void setFlag() const { _u.setFlag(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void print(std::ostream &os) const { os << "max("; _u.print(os); os<<")"; }
-private:
-    template<class U> inline
-    typename util::enable_if< util::is_same<U,gsFeSpace<Scalar> >::value, const gsMatrix<Scalar> & >::type
-    eval_impl(const U & u, const index_t k)  const
-    {
-        tmp = u.eval(k);
-
-        res.resize(1,u.cardinality());
-        if (E::ColBlocks)
-            for (index_t c=0; c!=_u.cardinality(); c++)
-                res(0,c) = tmp.block(0,c*u.cols(),u.rows(),u.cols()).maxCoeff();
-        else
-            for (index_t c=0; c!=_u.rows(); c++)
-                res(0,c) = tmp.block(c*u.rows(),0,u.rows(),u.cols()).maxCoeff();
-        return res;
-    }
-
-
-    template<class U> inline
-    typename util::enable_if< !util::is_same<U,gsFeSpace<Scalar> >::value, const gsMatrix<Scalar> & >::type
-    eval_impl(const U & u, const index_t k)  const
-    {
-        res = u.eval(k).colwise().maxCoeff();
-        return res;
-    }
-};
-
-template<class E>
-class rowsum_expr  : public _expr<rowsum_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E::Space, ColBlocks = E::ColBlocks};
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> tmp;
-
-public:
-
-    rowsum_expr(_expr<E> const& u) : _u(u)
-    {
-        //GISMO_ASSERT( _u.rows()*_u.cols() == _n*_m, "Wrong dimension"); //
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        tmp = _u.eval(k).rowwise().sum();
-        return tmp;
-    }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return 1; }
-    void setFlag() const { _u.setFlag(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void print(std::ostream &os) const { os << "rowsum("; _u.print(os); os<<")"; }
-};
-
-template<class E>
-class colsum_expr  : public _expr<colsum_expr<E> >
-{
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E::Space, ColBlocks = E::ColBlocks};
-private:
-    typename E::Nested_t _u;
-    mutable gsMatrix<Scalar> tmp;
-
-public:
-
-    colsum_expr(_expr<E> const& u) : _u(u)
-    {
-        //GISMO_ASSERT( _u.rows()*_u.cols() == _n*_m, "Wrong dimension"); //
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        tmp = _u.eval(k).colwise().sum();
-        return tmp;
-    }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return 1; }
-    void setFlag() const { _u.setFlag(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void print(std::ostream &os) const { os << "colsum("; _u.print(os); os<<")"; }
-};
-
-/**
-   Expression for the identity matrix
-*/
-class idMat_expr : public _expr<idMat_expr >
-{
-public:
-    typedef real_t Scalar;
-    enum {Space = 0, ScalarValued = 0, ColBlocks = 0};
-private:
-    index_t _dim;
-
-public:
-    idMat_expr(const index_t dim) : _dim(dim) { }
-
-public:
-
-    gsMatrix<Scalar>::IdentityReturnType eval(const index_t) const
-    {
-        return gsMatrix<Scalar>::Identity(_dim,_dim);
-    }
-
-    index_t rows() const { return _dim; }
-    index_t cols() const { return  _dim; }
-    void parse(gsExprHelper<Scalar> & ) const {  }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os << "id("<<_dim <<")";}
-};
-
-/**
-   Wrapper expression for constant matrices
-*/
-class constMat_expr : public _expr<constMat_expr >
-{
-public:
-    typedef real_t Scalar;
-    enum {Space = 0, ScalarValued = 0, ColBlocks = 0};
-private:
-    gsMatrix<Scalar> _mat;
-
-public:
-    constMat_expr(const gsMatrix<Scalar> mat) : _mat(mat) { }
-
-public:
-
-    gsMatrix<Scalar> eval(const index_t) const
-    {
-        return _mat;
-    }
-
-    index_t rows() const { return _mat.rows(); }
-    index_t cols() const { return  _mat.cols(); }
-    void parse(gsExprHelper<Scalar> & ) const {  }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os << "constMat";}
-};
-
-/**
-   Expression for the sign of another expression
-*/
-template<class E>
-class sign_expr : public _expr<sign_expr<E> >
-{
-    typename E::Nested_t _u;
-    typename E::Scalar _tol;
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 1, Space = E::Space, ColBlocks= 0};
-
-    sign_expr(_expr<E> const& u, Scalar tolerance = 0.0) : _u(u),_tol(tolerance){
-        GISMO_ASSERT( _tol >= 0, "Tolerance for sign_expr should be a positive number.");
-    }
-
-    Scalar eval(const index_t k) const
-    {
-        const Scalar v = _u.val().eval(k);
-        return ( v>_tol ? 1 : ( v<-_tol ? -1 : 0 ) );
-    }
-
-    static index_t rows() { return 0; }
-    static index_t cols() { return 0; }
-
-    void parse(gsExprHelper<Scalar> & el) const
-    { _u.parse(el); }
-
-    static bool isScalar() { return true; }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os<<"sgn("; _u.print(os); os <<")"; }
-};
-
-/**
-   Expression for the exponentiation of a given expression.
-*/
-template<class E>
-class exp_expr : public _expr<exp_expr<E> >
-{
-  typename E::Nested_t _u;
- public:
-  typedef typename E::Scalar Scalar;
-  enum {ScalarValued = 1, Space = E::Space, ColBlocks= 0};
-
-  exp_expr(_expr<E> const& u) : _u(u) { }
-
-  Scalar eval(const index_t k) const
-  {
-    const Scalar v = _u.val().eval(k);
-    return math::exp(v);
-  }
-
-  static index_t rows() { return 0; }
-  static index_t cols() { return 0; }
-
-  void parse(gsExprHelper<Scalar> & el) const
-  { _u.parse(el); }
-
-  static bool isScalar() { return true; }
-
-  const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-  const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-  void print(std::ostream &os) const { os<<"exp("; _u.print(os); os <<")"; }
-};
 
 /**
    Expression for the component-wise positive part
@@ -1300,41 +583,7 @@ class ppartval_expr : public _expr<ppartval_expr<E> >
   void print(std::ostream &os) const { os<<"posPart("; _u.print(os); os <<")"; }
 };
 
-/**
-   Expression pow(a,b) returns the value of 'a' raised to the power of 'b'
-*/
-template<class E>
-class pow_expr : public _expr<pow_expr<E> >
-{
-    typename E::Nested_t _u;
 
-public:
-    typedef typename E::Scalar Scalar;
-    enum {ScalarValued = 1, Space = E::Space, ColBlocks= E::ColBlocks};
-
-    Scalar _q;// power
-
-    pow_expr(_expr<E> const& u, Scalar q) : _u(u), _q(q) { }
-
-    Scalar eval(const index_t k) const
-    {
-        const Scalar v = _u.val().eval(k);
-        return math::pow(v,_q);
-    }
-
-    static index_t rows() { return 0; }
-    static index_t cols() { return 0; }
-
-    void parse(gsExprHelper<Scalar> & el) const
-    { _u.parse(el); }
-
-    static bool isScalar() { return true; }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os<<"pow("; _u.print(os); os <<")"; }
-};
 
 /**
    computes outer products of a matrix by a space of dimension > 1
@@ -1446,210 +695,6 @@ public:
     void print(std::ostream &os) const { os << "matrix_by_space_tr("; _u.print(os); os<<")"; }
 };
 
-/*
-  Adaptor for scalar-valued expression
-*/
-template<class E>
-class value_expr  : public _expr<value_expr<E> >
-{
-    typename E::Nested_t _u;
-
-public:
-    typedef typename E::Scalar Scalar;
-    value_expr(_expr<E> const& u) : _u(u)
-    {
-        // rows/cols not known at construction time
-        //GISMO_ASSERT(u.rows()*u.cols()<=1, "Expression\n"<<u<<"is not a scalar.");
-    }
-
-public:
-    enum {Space= 0, ScalarValued= 1, ColBlocks= 0};
-
-    Scalar eval(const index_t k) const { return eval_impl(_u,k); }
-
-    // enables expr.val().val()
-    inline value_expr<E> val() const { return *this; }
-    index_t rows() const { return 0; }
-    index_t cols() const { return 0; }
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    static bool isScalar() { return true; }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { _u.print(os); }
-
-    // Math functions. eg
-    // sqrt_mexpr<T> sqrt() { return sqrt_mexpr<T>(*this); }
-private:
-    template<class U> static inline
-    typename util::enable_if<U::ScalarValued,Scalar>::type
-    eval_impl(const U & u, const index_t k) { return u.eval(k); }
-
-    template<class U> static inline
-    typename util::enable_if<!U::ScalarValued,Scalar>::type
-    eval_impl(const U & u, const index_t k) { return u.eval(k).value(); }
-};
-
-template<class E>
-class abs_expr  : public _expr<abs_expr<E> >
-{
-    typename E::Nested_t _u;
-
-public:
-    typedef typename E::Scalar Scalar;
-    explicit abs_expr(_expr<E> const& u) : _u(u) { }
-
-public:
-    enum {Space= 0, ScalarValued= 1, ColBlocks= 0};
-
-    Scalar eval(const index_t k) const { return abs_expr::eval_impl(_u,k); }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-    static bool isScalar() { return true; }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { _u.print(os); }
-
-    // Math functions. eg
-    // sqrt_mexpr<T> sqrt() { return sqrt_mexpr<T>(*this); }
-private:
-    template<class U> static inline
-    typename util::enable_if<U::ScalarValued,Scalar>::type
-    eval_impl(const U & u, const index_t k) {return math::abs(u.eval(k)); }
-    template<class U> static inline
-    typename util::enable_if<!U::ScalarValued,gsMatrix<Scalar> >::type
-    eval_impl(const U & u, const index_t k) { return u.eval(k).cwiseAbs(); }
-};
-
-/*
-  Expression for the gradient of a finite element variable
-
-  Transposed gradient vectors are returned as a matrix
-*/
-template<class E>
-class grad_expr : public _expr<grad_expr<E> >
-{
-    typename E::Nested_t _u;
-public:
-    enum {Space = E::Space, ScalarValued= 0, ColBlocks= 0};
-
-    typedef typename E::Scalar Scalar;
-    mutable gsMatrix<Scalar> tmp;
-
-    grad_expr(const E & u) : _u(u)
-    { GISMO_ASSERT(1==u.dim(),"grad(.) requires 1D variable, use jac(.) instead.");}
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        // Assumes: derivatives are in _u.data().values[1]
-        // gsExprHelper acounts for compositions/physical expressions
-        // so that derivs are directly computed
-        tmp = _u.data().values[1].reshapeCol(k, cols(), cardinality_impl()).transpose();
-        return tmp;
-    }
-
-    index_t rows() const { return 1 /*==u.dim()*/; }
-
-    index_t cols() const { return _u.source().domainDim(); }
-
-    index_t cardinality_impl() const
-    { return _u.data().values[1].rows() / cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_u);
-        _u.data().flags |= NEED_GRAD;
-    }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const
-    {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os << "\u2207("; _u.print(os); os <<")"; }
-private:
-
-    template<class U> static inline
-    typename util::enable_if<util::is_same<U,gsComposition<Scalar> >::value,
-                             const gsMatrix<Scalar> &>::type
-    eval_impl(const U & u, const index_t k)
-    {
-        return u.eval(k);
-    }
-};
-
-/*
-  \brief Expression for the gradient of a finite element variable
-
-  Transposed gradient vectors are returned as a matrix.
-  This specialization is for a gsFeSolution object
-*/
-
-template<class T>
-class grad_expr<gsFeSolution<T> > : public _expr<grad_expr<gsFeSolution<T> > >
-{
-protected:
-    const gsFeSolution<T> _u;
-
-public:
-    typedef T Scalar;
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    explicit grad_expr(const gsFeSolution<T> & u) : _u(u) { }
-
-    mutable gsMatrix<T> res;
-    const gsMatrix<T> & eval(index_t k) const
-    {
-        GISMO_ASSERT(_u.check(), "Invalid state in gsFeSolution");
-        const gsDofMapper & map = _u.mapper();
-        auto & act = _u.data().actives.col(1 == _u.data().actives.cols() ? 0:k );
-        res.setZero(_u.dim(), _u.parDim());
-        for (index_t c = 0; c!= _u.dim(); c++)
-        {
-            for (index_t i = 0; i!=_u.data().actives.rows(); ++i)
-            {
-                const index_t ii = map.index(act[i], _u.data().patchId, c);
-                if ( map.is_free_index(ii) ) // DoF value is in the solVector
-                {
-                    res.row(c) += _u.coefs().at(ii) *
-                        _u.data().values[1].col(k).segment(i*_u.parDim(), _u.parDim()).transpose();
-                }
-                else
-                {
-                    res.row(c) +=
-                        _u.fixedPart().at( map.global_to_bindex(ii) ) *
-                        _u.data().values[1].col(k).segment(i*_u.parDim(), _u.parDim()).transpose();
-                }
-            }
-        }
-        return res;
-    }
-
-    index_t rows() const {return _u.dim();}
-    index_t cols() const {return _u.parDim(); }
-
-    const gsFeSpace<Scalar> & rowVar() const
-    {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const
-    {return gsNullExpr<Scalar>::get();}
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        _u.parse(evList);                         // add symbol
-        evList.add(_u.space());
-        _u.data().flags |= NEED_GRAD|NEED_ACTIVE; // define flags
-    }
-
-    void print(std::ostream &os) const { os << "\u2207(s)"; }
-};
 
 /*
   Expression for the derivative of the jacobian of a spline geometry map,
@@ -1703,325 +748,18 @@ public:
 };
 
 
-/*
-  Expression for the nabla (\f$\nabla\f$) of a finite element variable,
-*/
-template<class T>
-class nabla_expr : public _expr<nabla_expr<T> >
-{
-    typename gsFeVariable<T>::Nested_t u;
 
-public:
-    typedef T Scalar;
-    enum{Space = 1};
 
-    /* // todo
-       nabla_expr(const gsGeometryMap<T> & G)
-       : m_data(G.data()) { }
-    */
 
-    nabla_expr(const gsFeVariable<T> & _u) : u(_u)
-    {
-        //GISMO_ASSERT(u.parDim()==u.dim(),"nabla(.) requires tarDim==parDim:"
-        //             << u.parDim() <<"!="<< u.dim() <<"\n" );
-    }
-
-    mutable gsMatrix<Scalar> res;
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        const index_t d = u.cols();
-        const index_t n = rows() / d;
-        res.setZero(rows(), d);
-
-        for (index_t i = 0; i!=d; ++i)
-            res.col(i).segment(i*n,n) = u.data().values[1].reshapeCol(k, d, n).row(i);
-        return res;
-    }
-
-    index_t rows() const { return u.rows(); }
-    index_t cols() const { return u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(u);
-        u.data().flags |= NEED_GRAD;
-    }
-
-    const gsFeSpace<T> & rowVar() const { return u.rowVar(); }
-    const gsFeSpace<T> & colVar() const
-    {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os << "nabla("; u.print(os); os <<")"; }
-};
-
-/*
-  Expression for the nabla2 (\f$\nabla^2\f$ or Del2) of a finite element variable,
-  see also https://en.wikipedia.org/wiki/Del
-
-  Transposed pure second derivatives are returned as a matrix
-*/
-template<class T>
-class nabla2_expr : public _expr<nabla2_expr<T> >
-{
-    typename gsFeVariable<T>::Nested_t u;
-
-public:
-    typedef T Scalar;
-    enum{Space = 1};
-
-    /* // todo
-       nabla2_expr(const gsGeometryMap<T> & G)
-       : m_data(G.data()) { }
-    */
-
-    nabla2_expr(const gsFeVariable<T> & _u)
-    : u(_u)
-    { }
-
-    MatExprType eval(const index_t k) const
-    {
-        // numActive x parDim
-        return u.data().values[2]
-            .reShapeCol(k, u.data().values[2].rows()/u.cSize(), u.cSize() )
-            .topRows(u.parDim()).transpose();
-    }
-
-    index_t rows() const { return u.rows();   }
-    index_t cols() const { return u.parDim(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(u);
-        u.data().flags |= NEED_DERIV2;
-    }
-
-    const gsFeSpace<T> & rowVar() const { return u.rowVar(); }
-    const gsFeSpace<T> & colVar() const
-    {return gsNullExpr<T>::get();}
-};
 
 /// The nabla2 (\f$\nabla^2\f$) of a finite element variable
 template<class T>
 nabla2_expr<T> nabla2(const gsFeVariable<T> & u) { return nabla2_expr<T>(u); }
 // #define lapl(x) nabla2(x).sum() // assume tarDim==1
 
-/**
-   Expression for the outer pointing normal of a geometry map. This
-   expression is valid only at the boundaries of a geometric patch
-*/
-template<class T>
-class onormal_expr : public _expr<onormal_expr<T> >
-{
-    typename gsGeometryMap<T>::Nested_t _G;
 
-public:
-    typedef T Scalar;
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
 
-    explicit onormal_expr(const gsGeometryMap<T> & G) : _G(G) { }
 
-    auto eval(const index_t k) const -> decltype(_G.data().outNormals.col(k))
-    { return _G.data().outNormals.col(k); }
-
-    index_t rows() const { return  _G.source().targetDim(); }
-    index_t cols() const { return 1; }
-
-    const gsFeSpace<T> & rowVar() const {return gsNullExpr<T>::get();}
-    const gsFeSpace<T> & colVar() const {return gsNullExpr<T>::get();}
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_G);
-        _G.data().flags |= NEED_OUTER_NORMAL;
-    }
-
-    void print(std::ostream &os) const { os << "nv("; _G.print(os); os <<")"; }
-};
-
-/**
-   Expression for the out of plane surface normal of a geometry map.
-   The expression is valid for a surface or hypersurface.
-*/
-template<class T>
-class normal_expr : public _expr<normal_expr<T> >
-{
-    typename gsGeometryMap<T>::Nested_t _G;
-
-public:
-    typedef T Scalar;
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    normal_expr(const gsGeometryMap<T> & G) : _G(G)
-    {
-        GISMO_ENSURE( _G.source().domainDim()+1 == _G.source().targetDim(), "Surface normal requires codimension 1");
-    }
-
-    auto eval(const index_t k) const -> decltype(_G.data().normals.col(k))
-    { return _G.data().normals.col(k); }
-
-    index_t rows() const { return _G.source().targetDim(); }
-    index_t cols() const { return 1; }
-
-    const gsFeSpace<T> & rowVar() const {return gsNullExpr<T>::get();}
-    const gsFeSpace<T> & colVar() const {return gsNullExpr<T>::get();}
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_G);
-        _G.data().flags |= NEED_NORMAL;
-    }
-
-    void print(std::ostream &os) const { os << "sn("; _G.print(os); os <<")"; }
-};
-
-/**
-   Expression for the tangent vector of a geometry map. This
-   expression is valid only at the boundaries of a geometric patch
-*/
-template<class T>
-class tangent_expr : public _expr<tangent_expr<T> >
-{
-    typename gsGeometryMap<T>::Nested_t _G;
-
-public:
-    typedef T Scalar;
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    tangent_expr(const gsGeometryMap<T> & G) : _G(G) { }
-
-    mutable gsVector<Scalar> res;
-    const gsVector<Scalar> & eval(const index_t k) const
-    {
-        if (_G.targetDim()==2)
-        {
-            res = _G.data().outNormals.col(k);//2x1
-            std::swap( res(0,0), res(1,0) );
-            res(0,0) *= -1;
-            return res;
-        }
-        else if (_G.targetDim()==3)
-        {
-            res.resize(3);
-            res.col3d(0) = _G.data().normals.col3d(k)
-                .cross( _G.data().outNormals.col3d(k) );
-            return res;
-        }
-        else
-            GISMO_ERROR("Function not implemented for dimension"<<_G.targetDim());
-
-    }
-
-    index_t rows() const { return _G.source().targetDim(); }
-    index_t cols() const { return 1; }
-
-    static const gsFeSpace<Scalar> & rowVar() {return gsNullExpr<Scalar>::get();}
-    static const gsFeSpace<Scalar> & colVar() {return gsNullExpr<Scalar>::get();}
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_G);
-        _G.data().flags |= NEED_NORMAL;
-        _G.data().flags |= NEED_OUTER_NORMAL;
-    }
-
-    void print(std::ostream &os) const { os << "tv("; _G.print(os); os <<")"; }
-};
-
-/**
-   Expression for the Laplacian of a finite element variable
-*/
-template<class E>
-class lapl_expr : public _expr<lapl_expr<E> >
-{
-    typename E::Nested_t _u;
-
-public:
-    typedef typename E::Scalar Scalar;
-    enum {Space = E::Space, ScalarValued= 0, ColBlocks= 0};
-
-    lapl_expr(const E & u) : _u(u) { }
-
-    auto eval(const index_t k) const -> decltype(_u.data().laplacians.col(k))
-    {
-        // numActive x 1
-        return _u.data().laplacians.col(k);
-        //todo: replace by
-        // NEED_DERIV2
-        // ..nabla2.sum()
-    }
-
-    index_t rows() const { return _u.data().laplacians.rows(); }
-    index_t cols() const { return 1; }
-
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_u);
-        _u.data().flags |= NEED_LAPLACIAN;
-    }
-
-    static const gsFeSpace<Scalar> & rowVar() {return E::rowVar();}
-    static const gsFeSpace<Scalar> & colVar() {return gsNullExpr<Scalar>::get();}
-
-    void print(std::ostream &os) const { os << "\u2206("; _u.print(os); os <<")"; } //or \u0394
-};
-
-/*
-  Expression for the Laplacian of a finite element solution
-*/
-template<class T>
-class lapl_expr<gsFeSolution<T> > : public _expr<lapl_expr<gsFeSolution<T> > >
-{
-protected:
-    const gsFeSolution<T> _u;
-
-public:
-    typedef T Scalar;
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    lapl_expr(const gsFeSolution<T> & u) : _u(u) { }
-
-    mutable gsMatrix<T> res;
-    const gsMatrix<T> & eval(const index_t k) const
-    {
-        GISMO_ASSERT(_u.check(), "Invalid state in gsFeSolution");
-        const gsDofMapper & map = _u.mapper();
-        res.setZero(_u.dim(), 1); //  scalar, but per component
-
-        index_t numActs = _u.data().values[0].rows();
-        index_t numDers = _u.parDim() * (_u.parDim() + 1) / 2;
-        gsMatrix<T> deriv2;
-
-        auto & act = _u.data().actives.col(1 == _u.data().actives.cols() ? 0:k );
-        for (index_t c = 0; c!= _u.dim(); c++)
-            for (index_t i = 0; i!=numActs; ++i)
-            {
-                const index_t ii = map.index(act[i], _u.data().patchId, c);
-                deriv2 = _u.data().values[2].block(i*numDers,k,_u.parDim(),1); // this only takes d11, d22, d33 part. For all the derivatives [d11, d22, d33, d12, d13, d23]: col.block(i*numDers,k,numDers,1)
-                if ( map.is_free_index(ii) ) // DoF value is in the solVector
-                    res.at(c) += _u.coefs().at(ii) * deriv2.sum();
-                else
-                    res.at(c) +=_u.fixedPart().at( map.global_to_bindex(ii) ) * deriv2.sum();
-            }
-        return res;
-    }
-
-    index_t rows() const { return _u.dim(); }
-    index_t cols() const { return 1; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_u.space());
-        _u.data().flags |= NEED_ACTIVE | NEED_DERIV2;
-    }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<T>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<T>::get();}
-
-    void print(std::ostream &os) const { os << "\u2206(s)"; }
-};
 
 /*
   Expression for the (precomputed) second fundamental form of a surface
@@ -2437,258 +1175,6 @@ public:
     }
 };
 
-
-/*
-  Expression for the curl
-*/
-template<class T>
-class curl_expr : public _expr<curl_expr<T> >
-{
-public:
-    typedef T Scalar;
-private:
-    typename gsFeVariable<T>::Nested_t _u;
-    mutable gsMatrix<Scalar> res;
-public:
-    enum{ Space = 1, ScalarValued= 0, ColBlocks= 0};
-
-    curl_expr(const gsFeVariable<T> & u) : _u(u)
-    { GISMO_ASSERT(3==u.dim(),"curl(.) requires 3D variable."); }
-
-    const gsMatrix<T> & eval(const index_t k) const
-    {
-        res.setZero( rows(), _u.dim());
-        const index_t na = _u.data().values[0].rows();
-        gsAsConstMatrix<T, Dynamic, Dynamic> pd =
-            _u.data().values[1].reshapeCol(k, cols(), na);
-
-        res.col(0).segment(na  ,na) = -pd.row(2);
-        res.col(0).segment(2*na,na) =  pd.row(1);
-        res.col(1).segment(0   ,na) =  pd.row(2);
-        res.col(1).segment(2*na,na) = -pd.row(0);
-        res.col(2).segment(0   ,na) = -pd.row(1);
-        res.col(2).segment(na  ,na) =  pd.row(0);
-        return res;
-    }
-
-    index_t rows() const { return _u.dim() * _u.data().values[0].rows(); }
-    index_t cols() const { return _u.data().dim.first; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_u);
-        _u.data().flags |= NEED_GRAD;
-    }
-
-    const gsFeSpace<T> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<T> & colVar() const {return gsNullExpr<T>::get();}
-
-    void print(std::ostream &os) const { os << "curl("; _u.print(os); os <<")"; }
-};
-
-/*
-  Expression for multiplication operation (first version)
-
-  First argument E1 has ColBlocks = false
-
-  Partial specialization for (right) blockwise multiplication
-
-  B * [A1 A2 A3] = [B*A1  B*A2  B*A3]
-
-*/
-template <typename E1, typename E2>
-class mult_expr<E1,E2,false> : public _expr<mult_expr<E1, E2, false> >
-{
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-public:
-    enum {ScalarValued = E1::ScalarValued && E2::ScalarValued,
-        ColBlocks = E2::ColBlocks};
-    enum {Space = (int)E1::Space + (int)E2::Space };
-
-    typedef typename E1::Scalar Scalar;
-
-    mult_expr(_expr<E1> const& u,
-              _expr<E2> const& v)
-    : _u(u), _v(v) { }
-
-    mutable Temporary_t tmp;
-    const Temporary_t & eval(const index_t k) const
-    {
-        GISMO_ASSERT(0==_u.cols()*_v.rows() || _u.cols() == _v.rows(),
-                     "Wrong dimensions "<<_u.cols()<<"!="<<_v.rows()<<" in * operation:\n"
-                     << _u <<" times \n" << _v );
-
-        // Note: a * b * c --> (a*b).eval()*c
-        tmp = _u.eval(k) * _v.eval(k);
-        return tmp; // assumes result is not scalarvalued
-    }
-
-    index_t rows() const { return E1::ScalarValued ? _v.rows()  : _u.rows(); }
-    index_t cols() const { return E2::ScalarValued ? _u.cols()  : _v.cols(); }
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-
-    index_t cardinality_impl() const
-    { return 0==E1::Space ? _v.cardinality(): _u.cardinality(); }
-
-    const gsFeSpace<Scalar> & rowVar() const
-    { return 0==E1::Space ? _v.rowVar() : _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const
-    { return 0==E2::Space ? _u.colVar() : _v.colVar(); }
-
-    void print(std::ostream &os) const { _u.print(os); os<<"*"; _v.print(os); }
-};
-
-/*
-  Expression for multiplication operation (second version)
-
-  First argument E1 has ColBlocks = true
-
-  Partial specialization for (right) blockwise multiplication
-  [A1 A2 A3] * B = [A1*B  A2*B  A3*B]
-
-  as well as
-
-  both are ColBlocks: [A1 A2 A3] * [B1 B2 B3] = [A1*B1  A2*B2  A3*B3]
-                                                [A2*B1 ..           ]
-                                                [                   ]
-*/
-template <typename E1, typename E2>
-class mult_expr<E1, E2, true> : public _expr<mult_expr<E1, E2, true> >
-{
-public:
-    typedef typename E2::Scalar Scalar;
-private:
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-    mutable gsMatrix<Scalar> res;
-public:
-    enum {ScalarValued = 0, ColBlocks = E1::ColBlocks}; //(!)
-    enum {Space = (int)E1::Space + (int)E2::Space };
-
-    mult_expr(_expr<E1> const& u,
-              _expr<E2> const& v)
-    : _u(u), _v(v)
-    {
-
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const
-    {
-        const index_t uc = _u.cols();
-        const index_t ur = _u.rows();
-        const index_t nb = _u.cardinality();
-        const auto tmpA = _u.eval(k);
-        const auto tmpB = _v.eval(k);
-
-        const index_t vc = _v.cols();
-
-        // either _v.cardinality()==1 or _v.cardinality()==_u.cardinality()
-        if (  1 == _v.cardinality() ) //second is not ColBlocks
-        {
-            res.resize(ur, vc*nb);
-            GISMO_ASSERT(tmpA.cols()==uc*nb, "Dimension error.. "<< tmpA.cols()<<"!="<<uc*nb );
-            GISMO_ASSERT(1==_v.cardinality(), "Dimension error");
-            //gsInfo<<"cols = "<<res.cols()<<"; rows = "<<res.rows()<<"\n";
-            for (index_t i = 0; i!=nb; ++i)
-                res.middleCols(i*vc,vc).noalias()
-                    = tmpA.middleCols(i*uc,uc) * tmpB;
-        }
-        // both are ColBlocks: [A1 A2 A3] * [B1 B2 B3] = [A1*B1  A2*B2  A3*B3]
-        //                                               [A2*B1 ..           ]
-        //                                               [                   ]
-        else
-        {
-            const index_t nbv = _v.cardinality();
-            res.resize(ur*nb, vc*nbv);
-            for (index_t i = 0; i!=nb; ++i)
-                for (index_t j = 0; j!=nbv; ++j)
-                {
-                    res.block(i*ur,j*vc,ur,vc).noalias() =
-                        tmpA.middleCols(i*uc,uc) * tmpB.middleCols(j*vc,vc);
-                    // res.middleCols(i*vc,vc).noalias()
-                    //     = tmpA.middleCols(i*uc,uc) * tmpB.middleCols(i*vc,vc);
-                }
-        }
-        return res;
-    }
-
-    index_t rows() const {
-        return _u.rows();
-    }
-    index_t cols() const {
-        return _v.cols();
-    }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-
-    index_t cardinality_impl() const { return  _u.cardinality(); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const
-    {
-        if ( 1 == _v.cardinality() )
-            return _u.colVar();
-        else
-            return _v.colVar();
-    }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os);os <<"*";_v.print(os);os << ")"; }
-};
-
-/*
-  Expression for multiplication operation (third version)
-
-  Scalar multiplication
-*/
-template <typename E2>
-class mult_expr<typename E2::Scalar, E2, false>
-    : public _expr<mult_expr<typename E2::Scalar, E2, false> >
-// template <typename E> class scmult_expr : public _expr<scmult_expr<E> >
-{
-public:
-    typedef typename E2::Scalar Scalar;
-private:
-    Scalar const _c;
-    typename E2::Nested_t _v;
-
-    //mult_expr(const mult_expr&);
-public:
-    enum {ScalarValued = E2::ScalarValued, ColBlocks = E2::ColBlocks};
-    enum {Space = E2::Space};
-
-    mult_expr(Scalar const & c, _expr<E2> const& v)
-    : _c(c), _v(v) { }
-
-    EIGEN_STRONG_INLINE AutoReturn_t eval(const index_t k) const
-    {
-        return ( _c * _v.eval(k) );
-
-    }
-
-    index_t rows() const { return _v.rows(); }
-    index_t cols() const { return _v.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _v.parse(evList); }
-
-    index_t cardinality_impl() const
-    { return _v.cardinality(); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _v.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _v.colVar(); }
-
-    void print(std::ostream &os) const { os << _c <<"*";_v.print(os); }
-};
-
-
 template <typename E1, typename E2>
 class collapse_expr : public _expr<collapse_expr<E1, E2> >
 {
@@ -2760,305 +1246,6 @@ collapse_expr<E1,E2> collapse( _expr<E1> const& u, _expr<E2> const& v)
 { return collapse_expr<E1, E2>(u, v); }
 
 
-/*
-  Expression for the Frobenius matrix (or double dot) product (first
-  version) Also block-wise
-
-  [A1 A2 A3] . [B1 B2 B3]
-  =
-  [ A1.B1  A1.B2  A1.B3 ]
-  [ A2.B1  A2.B2  A2.B3 ]
-  [ A3.B1  A3.B2  A3.B3 ]
-*/
-template <typename E1, typename E2, bool = E2::ColBlocks>
-class frprod_expr : public _expr<frprod_expr<E1, E2> >
-{
-public:
-    typedef typename E1::Scalar Scalar;
-    enum {ScalarValued = 0, ColBlocks=E2::ColBlocks};
-    enum { Space = (int)E1::Space + (int)E2::Space };
-    // E1 E2 this (16 cases..)
-    // 0  0  0
-    // 1  1
-    // 2  2
-    // 3  3
-
-private:
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-    mutable gsMatrix<Scalar> res;
-
-public:
-
-    frprod_expr(_expr<E1> const& u, _expr<E2> const& v)
-    : _u(u), _v(v)
-    {
-        //todo: add check() functions, which will evaluate expressions on an empty matrix (no points) to setup initial dimensions ???
-        //GISMO_ASSERT(_u.rows() == _v.rows(),
-        //             "Wrong dimensions "<<_u.rows()<<"!="<<_v.rows()<<" in % operation");
-        //GISMO_ASSERT(_u.cols() == _v.cols(),
-        //             "Wrong dimensions "<<_u.cols()<<"!="<<_v.cols()<<" in % operation");
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const //todo: specialize for nb==1
-    {
-        // assert _u.size()==_v.size()
-        const index_t rb = _u.rows();
-        const index_t nb = _u.cardinality();
-        auto A = _u.eval(k);
-        auto B = _v.eval(k);
-        res.resize(nb, nb);
-        for (index_t i = 0; i!=nb; ++i) // all with all
-            for (index_t j = 0; j!=nb; ++j)
-                res(i,j) =
-                    (A.middleCols(i*rb,rb).array() * B.middleCols(j*rb,rb).array()).sum();
-        return res;
-    }
-
-    index_t rows() const { return _u.cols() / _u.rows(); }
-    index_t cols() const { return _u.cols() / _u.rows(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _v.colVar(); }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os); os<<" % "; _v.print(os); os<<")";}
-};
-
-/*
-
-  Expression for the Frobenius matrix (or double dot) product (second
-  version), When left hand only side is block-wise
-
-  [A1 A2 A3] : B = [A1:B  A2:B  A3:B]
-*/
-template <typename E1, typename E2>
-class frprod_expr<E1,E2,false> : public _expr<frprod_expr<E1, E2,false> >
-{
-public:
-    typedef typename E1::Scalar Scalar;
-    enum {ScalarValued = 0, Space = E1::Space, ColBlocks= E1::ColBlocks};
-
-private:
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-    mutable gsMatrix<Scalar> res;
-
-public:
-
-    frprod_expr(_expr<E1> const& u, _expr<E2> const& v)
-    : _u(u), _v(v)
-    {
-        // gsInfo << "expression is space ? "<<E1::Space <<"\n"; _u.print(gsInfo);
-        // GISMO_ASSERT(_u.rows() == _v.rows(),
-        //              "Wrong dimensions "<<_u.rows()<<"!="<<_v.rows()<<" in % operation");
-        // GISMO_ASSERT(_u.cols() == _v.cols(),
-        //              "Wrong dimensions "<<_u.cols()<<"!="<<_v.cols()<<" in % operation");
-    }
-
-    const gsMatrix<Scalar> & eval(const index_t k) const //todo: specialize for nb==1
-    {
-        // assert _u.size()==_v.size()
-        auto A = _u.eval(k);
-        auto B = _v.eval(k);
-        const index_t rb = A.rows(); //==cb
-        const index_t nb = _u.cardinality();
-        res.resize(nb, 1);
-        for (index_t i = 0; i!=nb; ++i) // all with all
-            res(i,0) =
-                (A.middleCols(i*rb,rb).array() * B.array()).sum();
-        return res;
-    }
-
-    index_t rows() const { return _u.cols() / _u.rows(); }
-    index_t cols() const { return 1; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _v.rowVar(); }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os); os<<" % "; _v.print(os); os<<")";}
-};
-
-/*
-  Expression for scalar division operation (first version)
-*/
-template <typename E1, typename E2>
-class divide_expr : public _expr<divide_expr<E1,E2> >
-{
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-public:
-    typedef typename E1::Scalar Scalar;
-
-public:
-    enum {ScalarValued = E1::ScalarValued, ColBlocks= E2::ColBlocks};
-    enum {Space = E1::Space}; // The denominator E2 has to be scalar.
-
-    divide_expr(_expr<E1> const& u, _expr<E2> const& v)
-    : _u(u), _v(v)
-    {
-        GISMO_STATIC_ASSERT(E2::ScalarValued, "The denominator needs to be scalar valued.");
-    }
-
-    AutoReturn_t eval(const index_t k) const
-    { return ( _u.eval(k) / _v.eval(k) ); }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os);os <<" / ";_v.print(os);os << ")"; }
-};
-
-/*
-  Division specialization (second version) for constant value denominator
-*/
-template <typename E1>
-class divide_expr<E1,typename E1::Scalar>
-    : public _expr<divide_expr<E1,typename E1::Scalar> >
-{
-public:
-    typedef typename E1::Scalar Scalar;
-
-private:
-    typename E1::Nested_t _u;
-    Scalar  const   _c;
-
-public:
-    enum {Space= E1::Space, ScalarValued = E1::ScalarValued, ColBlocks= E1::ColBlocks};
-
-    divide_expr(_expr<E1> const& u, Scalar const  c)
-    : _u(u), _c(c) { }
-
-    AutoReturn_t eval(const index_t k) const
-    { return ( _u.eval(k) / _c ); }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os);os <<"/"<< _c << ")"; }
-};
-
-/*
-  Division specialization (third version) for constant value
-  numerator
-*/
-template <typename E2>
-class divide_expr<typename E2::Scalar,E2>
-    : public _expr<divide_expr<typename E2::Scalar,E2> >
-{
-public:
-    typedef typename E2::Scalar Scalar;
-
-private:
-    Scalar  const   _c;
-    typename E2::Nested_t _u;
-public:
-    enum {Space= 0, ScalarValued = 1, ColBlocks= 0};
-
-    divide_expr(Scalar const c, _expr<E2> const& u)
-    : _c(c), _u(u)
-    { GISMO_STATIC_ASSERT(E2::ScalarValued, "The denominator needs to be scalar valued."); }
-
-    Scalar eval(const index_t k) const
-    { return ( _c / _u.val().eval(k) ); }
-
-    index_t rows() const { return 0; }
-    index_t cols() const { return 0; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); }
-
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _u.colVar(); }
-
-    void print(std::ostream &os) const
-    { os << "("<< _c <<"/";_u.print(os);os << ")";}
-};
-
-/*
-  Expression for addition operation
-*/
-template <typename E1, typename E2>
-class add_expr : public _expr<add_expr<E1, E2> >
-{
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-public:
-    enum {ScalarValued = E1::ScalarValued && E2::ScalarValued,
-        ColBlocks = E1::ColBlocks || E2::ColBlocks };
-    enum {Space = E1::Space}; // == E2::Space
-
-    typedef typename E1::Scalar Scalar;
-
-    add_expr(_expr<E1> const& u, _expr<E2> const& v)
-    : _u(u), _v(v)
-    {
-        GISMO_ENSURE((int)E1::Space == (int)E2::Space &&
-                     _u.rowVar()==_v.rowVar() && _u.colVar()==_v.colVar(),
-                     "Error: adding apples and oranges (use comma instead),"
-                     " namely:\n" << _u <<"\n"<<_v<<
-                     " \nvars:\n" << _u.rowVar().id()<<"!="<<_v.rowVar().id() <<", "<< _u.colVar().id()<<"!="<<_v.colVar().id()<<
-                     " \nspaces:\n" << (int)E1::Space<< "!="<< (int)E2::Space
-            );
-    }
-
-    mutable Temporary_t res;
-    const Temporary_t & eval(const index_t k) const
-    {
-        GISMO_ASSERT(_u.rows() == _v.rows(),
-                     "Wrong dimensions "<<_u.rows()<<"!="<<_v.rows()<<" in + operation:\n"
-                     << _u <<" plus \n" << _v );
-        GISMO_ASSERT(_u.cols() == _v.cols(),
-                     "Wrong dimensions "<<_u.cols()<<"!="<<_v.cols()<<" in + operation:\n"
-                     << _u <<" plus \n" << _v );
-        res = _u.eval(k) + _v.eval(k);
-        return res;
-    }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-
-    index_t cardinality_impl() const { return _u.cardinality_impl(); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _v.colVar(); }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os);os <<" + ";_v.print(os);os << ")"; }
-};
 
 /*
   lincom_expr (lc) ?
@@ -3121,71 +1308,6 @@ private:
     typename E2::Nested_t _M;
 
     mutable gsMatrix<Scalar> res;
-};
-
-
-/*
-  Expression for subtraction operation
-*/
-template <typename E1, typename E2>
-class sub_expr : public _expr<sub_expr<E1, E2> >
-{
-    typename E1::Nested_t _u;
-    typename E2::Nested_t _v;
-
-public:
-    enum {ScalarValued = E1::ScalarValued && E2::ScalarValued,
-        ColBlocks = E1::ColBlocks || E2::ColBlocks };
-    enum {Space = E1::Space}; // == E2::Space
-
-    typedef typename E1::Scalar Scalar;
-
-    sub_expr(_expr<E1> const& u, _expr<E2> const& v)
-    : _u(u), _v(v)
-    {
-        GISMO_ENSURE((int)E1::Space == (int)E2::Space &&
-                     _u.rowVar()==_v.rowVar() && _u.colVar()==_v.colVar(),
-                     "Error: substracting apples from oranges (use comma instead),"
-                     " namely:\n" << _u <<"\n"<<_v);
-    }
-
-    mutable Temporary_t res;
-    const Temporary_t & eval(const index_t k) const
-    {
-        // GISMO_ASSERT(_u.rowVar().id()==_v.rowVar().id() && _u.rowVar().isAcross()==_v.rowVar().isAcross(),
-        //     "The row spaces are not split compatibly.");
-        // GISMO_ASSERT(_u.colVar().id()==_v.colVar().id() && _u.colVar().isAcross()==_v.colVar().isAcross(),
-        //     "The col spaces are not split compatibly.");
-        GISMO_ASSERT(_u.rows() == _v.rows(),
-                     "Wrong dimensions "<<_u.rows()<<"!="<<_v.rows()<<" in - operation:\n" << _u <<" minus \n" << _v );
-        GISMO_ASSERT(_u.cols() == _v.cols(),
-                     "Wrong dimensions "<<_u.cols()<<"!="<<_v.cols()<<" in - operation:\n" << _u <<" minus \n" << _v );
-        GISMO_ASSERT(_u.cardinality() == _u.cardinality(),
-                     "Cardinality "<< _u.cardinality()<<" != "<< _v.cardinality());
-        //return (_u.eval(k) - _v.eval(k) ).eval();
-        //return (_u.eval(k) - _v.eval(k) ); // any temporary matrices eval(.) will leak mem.
-        res = _u.eval(k) - _v.eval(k);
-        return res;
-    }
-
-    index_t rows() const { return _u.rows(); }
-    index_t cols() const { return _u.cols(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    { _u.parse(evList); _v.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _u.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _v.colVar(); }
-
-    index_t cardinality_impl() const
-    {
-        GISMO_ASSERT(_u.cardinality() == _u.cardinality(),
-                     "Cardinality "<< _u.cardinality()<<" != "<< _v.cardinality());
-        return _u.cardinality();
-    }
-
-    void print(std::ostream &os) const
-    { os << "("; _u.print(os); os<<" - ";_v.print(os); os << ")";}
 };
 
 /*
@@ -3271,17 +1393,37 @@ public:
 /// The identity matrix of dimension \a dim
 EIGEN_STRONG_INLINE idMat_expr id(const index_t dim) { return idMat_expr(dim); }
 
-EIGEN_STRONG_INLINE constMat_expr ones(const index_t dim) {
+EIGEN_STRONG_INLINE constMat_expr ones(const index_t dim)
+{
     gsMatrix<real_t> ones(dim, dim);
     ones.fill(1);
     return constMat_expr(ones);
-    }
+}
 
 EIGEN_STRONG_INLINE constMat_expr mat(const gsMatrix<real_t> mat) { return constMat_expr(mat); }
 
 // Returns the unit as an expression
 //EIGEN_STRONG_INLINE _expr<real_t> one() { return _expr<real_t,true>(1); }
 
+/// Replicate an expression
+template <typename E> EIGEN_STRONG_INLINE
+replicate_expr<E> const replicate(E const & u, index_t n, index_t m = 1)
+{ return replicate_expr<E>(u, n, m); }
+
+/// Reshape an expression
+template <class E> EIGEN_STRONG_INLINE
+reshape_expr<E> const reshape(E const & u, index_t n, index_t m)
+{ return reshape_expr<E>(u, n, m); }
+
+/// Make a matrix 2x2 expression "flat"
+template <typename E> EIGEN_STRONG_INLINE
+flat_expr<E> const flat(E const & u)
+{ return flat_expr<E>(u); }
+
+/// Get diagonal elements of matrix as a vector
+template <typename E> EIGEN_STRONG_INLINE
+diag_expr<E> const diagonal(E const & u)
+{ return diag_expr<E>(u); }
 
 /// Absolute value
 template<class E> EIGEN_STRONG_INLINE
